refactor: split stamp service boundary
