+++ backend/app/repositories/page_repository.py
@@ -4,10 +4,8 @@ from ..models import CategoryFilter, CourseMood
 from ..repository_normalized import (
     get_bootstrap,
     get_place,
-    get_stamps,
     list_courses,
     list_places,
-    toggle_stamp,
 )
 
 
@@ -25,18 +23,3 @@ def read_place_entry(db: Session, place_id: str):
 
 def list_course_entries(db: Session, mood: CourseMood | None):
     return list_courses(db, mood)
-
-
-def read_stamp_state(db: Session, user_id: str | None):
-    return get_stamps(db, user_id)
-
-
-def toggle_stamp_entry(
-    db: Session,
-    user_id: str,
-    place_id: str,
-    latitude: float,
-    longitude: float,
-    unlock_radius_meters: int,
-):
-    return toggle_stamp(db, user_id, place_id, latitude, longitude, unlock_radius_meters)
