+++ backend/tests/test_page_service.py
@@ -3,20 +3,7 @@ from types import SimpleNamespace
 import pytest
 from fastapi import HTTPException, status
 
-from app.models import SessionUser, StampToggleRequest
 from app.services import page_service
-
-
-def build_session_user() -> SessionUser:
-    return SessionUser(
-        id="user-1",
-        nickname="tester",
-        email="tester@example.com",
-        provider="kakao",
-        profileImage=None,
-        isAdmin=False,
-        profileCompletedAt=None,
-    )
 
 
 def test_read_place_service_maps_missing_place_to_404(monkeypatch):
@@ -27,39 +14,5 @@ def test_read_place_service_maps_missing_place_to_404(monkeypatch):
 
     with pytest.raises(HTTPException) as caught:
         page_service.read_place_service(SimpleNamespace(), "missing-place")
-
-    assert caught.value.status_code == status.HTTP_404_NOT_FOUND
-
-
-def test_toggle_stamp_service_maps_permission_error_to_403(monkeypatch):
-    def failing_toggle_stamp(*_args, **_kwargs):
-        raise PermissionError("forbidden")
-
-    monkeypatch.setattr(page_service, "toggle_stamp_entry", failing_toggle_stamp)
-
-    with pytest.raises(HTTPException) as caught:
-        page_service.toggle_stamp_service(
-            SimpleNamespace(),
-            StampToggleRequest(placeId="place-1", latitude=36.35, longitude=127.38),
-            build_session_user(),
-            page_service.Settings(stamp_unlock_radius_meters=120),
-        )
-
-    assert caught.value.status_code == status.HTTP_403_FORBIDDEN
-
-
-def test_toggle_stamp_service_maps_missing_place_to_404(monkeypatch):
-    def failing_toggle_stamp(*_args, **_kwargs):
-        raise ValueError("장소를 찾을 수 없어요.")
-
-    monkeypatch.setattr(page_service, "toggle_stamp_entry", failing_toggle_stamp)
-
-    with pytest.raises(HTTPException) as caught:
-        page_service.toggle_stamp_service(
-            SimpleNamespace(),
-            StampToggleRequest(placeId="missing-place", latitude=36.35, longitude=127.38),
-            build_session_user(),
-            page_service.Settings(stamp_unlock_radius_meters=120),
-        )
 
     assert caught.value.status_code == status.HTTP_404_NOT_FOUND
